# swagger/admin-api.yaml
openapi: 3.0.3
info:
  title: API Absensi - Admin & Super Admin Endpoints
  description: |
    Dokumentasi API untuk endpoints Admin dan Super Admin sistem absensi.
    
    **Level Akses:**
    - Super Admin (Level 1): Akses penuh ke semua endpoint
    - Admin OPD (Level 2): Akses terbatas ke SKPD yang ditugaskan
    - Admin UPT (Level 3): Akses terbatas ke Satker yang ditugaskan
    
    **Autentikasi:** Semua endpoint memerlukan JWT token dengan level admin yang sesuai.
  version: 1.0.0
  contact:
    name: API Support
    email: support@absensi.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/admin
    description: Development Server
  - url: https://api.absensi.com/api/admin
    description: Production Server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token untuk autentikasi admin
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: string
          example: "Detailed error description"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin001"
        full_name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "admin@example.com"
        level:
          type: string
          enum: ["1", "2", "3", "4"]
          example: "1"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /dashboard/overview:
    get:
      tags:
        - Dashboard
      summary: Get dashboard overview statistics
      description: Mendapatkan statistik overview dashboard untuk super admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard overview data
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUsers:
                    type: integer
                    example: 150
                  totalLocations:
                    type: integer
                    example: 25
                  todayAttendance:
                    type: integer
                    example: 120
                  attendanceRate:
                    type: number
                    format: float
                    example: 80.5
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dashboard/weekly-stats:
    get:
      tags:
        - Dashboard
      summary: Get weekly attendance statistics
      description: Mendapatkan statistik kehadiran mingguan
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Weekly statistics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  weeklyData:
                    type: array
                    items:
                      type: object
                      properties:
                        day:
                          type: string
                          example: "Monday"
                        present:
                          type: integer
                          example: 120
                        absent:
                          type: integer
                          example: 30
  /dashboard/top-users:
    get:
      tags:
        - Dashboard
      summary: Get top attendance users
      description: Mendapatkan daftar pengguna dengan kehadiran tertinggi
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Jumlah pengguna yang ditampilkan
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Top users data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    user_id:
                      type: integer
                      example: 123
                    full_name:
                      type: string
                      example: "John Doe"
                    attendance_count:
                      type: integer
                      example: 25
                    attendance_rate:
                      type: number
                      format: float
                      example: 96.2
  /dashboard/location-stats:
    get:
      tags:
        - Dashboard
      summary: Get location attendance statistics
      description: Mendapatkan statistik kehadiran per lokasi
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Location statistics data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    lokasi_id:
                      type: integer
                      example: 1
                    nama_lokasi:
                      type: string
                      example: "Kantor Pusat"
                    total_attendance:
                      type: integer
                      example: 500
                    today_attendance:
                      type: integer
                      example: 25
  /dashboard/realtime:
    get:
      tags:
        - Dashboard
      summary: Get real-time attendance stats
      description: Mendapatkan statistik kehadiran real-time hari ini
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Real-time statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_expected:
                    type: integer
                    example: 150
                  present:
                    type: integer
                    example: 120
                  absent:
                    type: integer
                    example: 30
                  on_leave:
                    type: integer
                    example: 5
                  late:
                    type: integer
                    example: 15

  # User Management Endpoints
  /register:
    post:
      tags:
        - User Management
      summary: Register new user
      description: Mendaftarkan pengguna baru (Super Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - full_name
                - email
                - level
              properties:
                username:
                  type: string
                  example: "admin001"
                password:
                  type: string
                  example: "securepassword"
                full_name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  example: "admin@example.com"
                level:
                  type: string
                  enum: ["1", "2", "3", "4"]
                  example: "2"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      tags:
        - User Management
      summary: Get all users
      description: Mendapatkan daftar semua pengguna dengan pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Jumlah data per halaman
          schema:
            type: integer
            default: 10
        - name: level
          in: query
          description: Filter berdasarkan level pengguna
          schema:
            type: string
            enum: ["1", "2", "3", "4"]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 100
                      totalPages:
                        type: integer
                        example: 10
...
  /users/{id}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      description: Mendapatkan detail pengguna berdasarkan ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - User Management
      summary: Update user
      description: Mengupdate data pengguna
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  example: "John Doe Updated"
                email:
                  type: string
                  example: "newemail@example.com"
                is_active:
                  type: boolean
                  example: true
                level:
                  type: string
                  enum: ["1", "2", "3", "4"]
                  example: "2"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}/force-logout:
    post:
      tags:
        - User Management
      summary: Force logout user
      description: Memaksa logout pengguna dengan menghapus refresh token
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User has been logged out successfully"

  # Location Management Endpoints
  /lokasi:
    post:
      tags:
        - Location Management
      summary: Create new location
      description: Membuat lokasi baru untuk presensi
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nama_lokasi
                - alamat
                - latitude
                - longitude
                - radius
              properties:
                nama_lokasi:
                  type: string
                  example: "Kantor Pusat"
                alamat:
                  type: string
                  example: "Jl. Sudirman No. 1"
                latitude:
                  type: number
                  format: float
                  example: -6.1944
                longitude:
                  type: number
                  format: float
                  example: 106.8229
                radius:
                  type: integer
                  example: 100
                id_skpd:
                  type: string
                  example: "01"
                id_satker:
                  type: string
                  example: "001"
                id_bidang:
                  type: string
                  example: "001"
      responses:
        '201':
          description: Location created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lokasi'

    get:
      tags:
        - Location Management
      summary: Get all locations
      description: Mendapatkan daftar semua lokasi
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Jumlah data per halaman
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lokasi'
                  pagination:
                    type: object
...
  /lokasi/{lokasi_id}:
    get:
      tags:
        - Location Management
      summary: Get location by ID
      description: Mendapatkan detail lokasi berdasarkan ID
      security:
        - bearerAuth: []
      parameters:
        - name: lokasi_id
          in: path
          required: true
          description: Location ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Location details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lokasi'

    patch:
      tags:
        - Location Management
      summary: Update location
      description: Mengupdate data lokasi
      security:
        - bearerAuth: []
      parameters:
        - name: lokasi_id
          in: path
          required: true
          description: Location ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nama_lokasi:
                  type: string
                  example: "Kantor Pusat Updated"
                alamat:
                  type: string
                  example: "Jl. Sudirman No. 2"
                latitude:
                  type: number
                  format: float
                  example: -6.1945
                longitude:
                  type: number
                  format: float
                  example: 106.8230
                radius:
                  type: integer
                  example: 150
      responses:
        '200':
          description: Location updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lokasi'

    delete:
      tags:
        - Location Management
      summary: Delete location
      description: Menghapus lokasi
      security:
        - bearerAuth: []
      parameters:
        - name: lokasi_id
          in: path
          required: true
          description: Location ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Location deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Location deleted successfully"

  # Attendance Management Endpoints
  /kehadiran:
    get:
      tags:
        - Attendance Management
      summary: Get all attendance records
      description: Mendapatkan semua record kehadiran dengan filter
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Nomor halaman
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Jumlah data per halaman
          schema:
            type: integer
            default: 10
        - name: start_date
          in: query
          description: Tanggal mulai (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: end_date
          in: query
          description: Tanggal akhir (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-31"
        - name: status
          in: query
          description: Filter berdasarkan status kehadiran
          schema:
            type: string
            enum: ["HADIR", "izin", "sakit", "cuti", "dinas"]
      responses:
        '200':
          description: List of attendance records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Kehadiran'
                  pagination:
                    type: object
  /kehadiran/{id}:
    get:
      tags:
        - Attendance Management
      summary: Get attendance by ID
      description: Mendapatkan detail kehadiran berdasarkan ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Attendance ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Attendance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Kehadiran'
  /kehadiran/user/{user_id}:
    get:
      tags:
        - Attendance Management
      summary: Get attendance by user ID
      description: Mendapatkan kehadiran berdasarkan user ID
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
            example: 123
        - name: start_date
          in: query
          description: Tanggal mulai
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: Tanggal akhir
          schema:
            type: string
            format: date
      responses:
        '200':
          description: User attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Kehadiran'
...
  /kehadiran/export/harian:
    get:
      tags:
        - Export
      summary: Export daily attendance
      description: Export data presensi harian ke Excel
      security:
        - bearerAuth: []
      parameters:
        - name: tanggal
          in: query
          required: true
          description: Tanggal export (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: skpd
          in: query
          description: Filter SKPD
          schema:
            type: string
        - name: lokasi
          in: query
          description: Filter lokasi
          schema:
            type: integer
      responses:
        '200':
          description: Excel file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
...